% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traveltimes-package.R
\docType{package}
\name{traveltimes-package}
\alias{traveltimes-package}
\alias{_PACKAGE}
\title{traveltimes: Urban travel time matrices using any and all modes of transport}
\description{
Urban travel time matrices using any and all modes of transport.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UrbanAnalyst/traveltimes}
  \item Report bugs at \url{https://github.com/UrbanAnalyst/traveltimes/issues}
}

}
\author{
\strong{Maintainer}: Mark Padgham \email{mark.padgham@email.com}

}
\keyword{internal}
